requirements.txt
#Django==1.11.6
django-extensions #==1.9.6
six==1.11.0
future==0.15.2
mock==1.0.1
nose==1.3.3
pep8==1.5.7
pyflakes==0.8.1
